move string format to details/helper
