+++ src/details/r3d_helper.h
@@ -10,6 +10,8 @@
 #define R3D_DETAILS_MACROS_H
 
 #include <stdint.h>
+#include <stdarg.h>
+#include <stdio.h>
 
 // ========================================
 // HELPER MACROS
@@ -34,6 +36,18 @@
 // ========================================
 // HELPER FUNCTIONS
 // ========================================
+
+static inline void r3d_string_format(char *dst, size_t dstSize, const char *fmt, ...)
+{
+    va_list args;
+    va_start(args, fmt);
+    int written = vsnprintf(dst, dstSize, fmt, args);
+    va_end(args);
+
+    if (written < 0 || (size_t)written >= dstSize) {
+        dst[dstSize - 1] = '\0';
+    }
+}
 
 static inline int r3d_lsb_index(uint32_t value)
 {
